Switch to ESModule imports / exports
