+++ packages/gatsby-source-nacelle/gatsby-node-es6.js
@@ -1,0 +1,260 @@
+import { ensureDir } from 'fs-extra';
+import { print } from 'gatsby/graphql';
+import {
+  sourceAllNodes,
+  createSchemaCustomization,
+  compileNodeQueries,
+  readOrGenerateDefaultFragments,
+  buildNodeDefinitions,
+  loadSchema,
+  createDefaultQueryExecutor
+} from 'gatsby-graphql-source-toolkit';
+
+import dotenv from 'dotenv';
+dotenv.config();
+
+const CHUNK_SIZE = 100;
+const fragmentsDir = process.cwd() + `/gql-fragments`;
+let cmsPreviewActive = true;
+
+const PaginateNacelle = {
+  name: 'NacellePagination',
+  expectedVariableNames: ['first', 'after'],
+  start() {
+    return {
+      variables: { first: CHUNK_SIZE, after: undefined },
+      hasNextPage: true
+    };
+  },
+  next(state, page) {
+    return {
+      variables: { first: CHUNK_SIZE, after: page.nextToken },
+      hasNextPage: Boolean(page.nextToken && page.items.length === CHUNK_SIZE)
+    };
+  },
+  concat(acc, page) {
+    return {
+      ...acc,
+      items: {
+        ...acc.items,
+        ...page.items
+      },
+      nextToken: page.nextToken
+    };
+  },
+  getItems(pageOrResult) {
+    return pageOrResult.items;
+  }
+};
+
+async function createSourcingConfig(gatsbyApi, pluginOptions) {
+  const { verbose } = pluginOptions;
+
+  function getPluginOption(option) {
+    // Accepts any camelCase, PascalCase, kebab-case, or snake_case option
+    const sanitize = (str) => str.replace(/[-_]/g, '').toLowerCase();
+    return pluginOptions[
+      Object.keys(pluginOptions).find(
+        (key) => sanitize(key) === sanitize(option)
+      )
+    ];
+  }
+
+  cmsPreviewActive = getPluginOption('cmsPreview') !== false;
+
+  // Extract credentials from plugin options
+  const nacelleSpaceId = getPluginOption('nacellespaceid');
+  if (!nacelleSpaceId) {
+    throw new Error(`Please provide a Nacelle Space ID to 'gatsby-source-nacelle'. For example:
+
+      {
+        resolve: 'gatsby-source-nacelle',
+        options: {
+          nacelleSpaceId: process.env.NACELLE_SPACE_ID,
+          nacelleGraphqlToken: process.env.NACELLE_GRAPHQL_TOKEN
+        }
+      }
+    `);
+  }
+
+  const nacelleGraphqlToken = getPluginOption('nacellegraphqltoken');
+  if (!nacelleGraphqlToken) {
+    throw new Error(`Please provide a Nacelle GraphQL Token to 'gatsby-source-nacelle'. For example:
+
+        {
+          resolve: 'gatsby-source-nacelle',
+          options: {
+            nacelleSpaceId: process.env.NACELLE_SPACE_ID,
+            nacelleGraphqlToken: process.env.NACELLE_GRAPHQL_TOKEN
+          }
+        }
+      `);
+  }
+
+  // Set up remote schema:
+  const defaultExecute = createDefaultQueryExecutor(
+    'https://hailfrequency.com/v2/graphql',
+    {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+        'x-nacelle-space-id': nacelleSpaceId,
+        'x-nacelle-space-token': nacelleGraphqlToken
+      }
+    }
+  );
+
+  const execute = (args) => {
+    if (verbose) {
+      console.log(args.operationName, args.variables);
+    }
+
+    return defaultExecute(args);
+  };
+
+  const schema = await loadSchema(execute);
+
+  // Configure Gatsby node types
+  // Note:
+  //  Queries with LIST_ prefix are automatically executed in sourceAllNodes
+  //  Queries with NODE_ prefix are executed in sourceNodeChanges for updated nodes
+  const gatsbyNodeTypes = [
+    {
+      remoteTypeName: 'Product',
+      queries: `
+        query LIST_PRODUCTS {
+          getProducts(first: $first, after: $after) {
+            nextToken
+            items { ..._ProductId_ }
+          }
+        }
+        query NODE_PRODUCT {
+          getProductByHandle(handle: $handle, locale: $locale) {
+            ..._ProductId_
+          }
+        }
+        fragment _ProductId_ on Product { __typename handle locale }
+      `
+    },
+    {
+      remoteTypeName: 'Collection',
+      queries: `
+        query LIST_COLLECTION {
+          getCollections(first: $first, after: $after) {
+            nextToken
+            items { ..._CollectionId_ }
+          }
+        }
+        query NODE_COLLECTION {
+          getCollectionByHandle(handle: $handle, locale: $locale) {
+            ..._CollectionId_
+          }
+        }
+        fragment _CollectionId_ on Collection { __typename handle locale }
+      `
+    },
+    {
+      remoteTypeName: 'Content',
+      queries: `
+        query LIST_CONTENT {
+          getContent(first: $first, after: $after) {
+            nextToken
+            items { ..._ContentId_ }
+          }
+        }
+        query NODE_CONTENT {
+          getContentByHandle(
+            type: $type
+            handle: $handle
+            locale: $locale
+          ) {
+            ..._ContentId_
+          }
+        }
+        fragment _ContentId_ on Content {
+          __typename
+          type
+          handle
+          locale
+        }
+      `
+    },
+    {
+      remoteTypeName: 'Space',
+      queries: `
+        query NODE_SPACE {
+          getSpace {
+            ..._SpaceId_
+          }
+        }
+        fragment _SpaceId_ on Space { __typename id }
+      `
+    }
+  ];
+
+  const gatsbyNodeTypesLessContent = gatsbyNodeTypes.filter(
+    (nodeType) => nodeType.remoteTypeName !== 'Content'
+  );
+
+  function getNodeTypes() {
+    return cmsPreviewActive ? gatsbyNodeTypesLessContent : gatsbyNodeTypes;
+  }
+
+  // Provide (or generate) fragments with fields to be fetched
+  ensureDir(fragmentsDir);
+  const fragments = await readOrGenerateDefaultFragments(fragmentsDir, {
+    schema,
+    gatsbyNodeTypes: getNodeTypes()
+  });
+
+  // Compile sourcing queries
+  const documents = compileNodeQueries({
+    schema,
+    gatsbyNodeTypes: getNodeTypes(),
+    customFragments: fragments
+  });
+
+  return {
+    gatsbyApi,
+    schema,
+    execute,
+    gatsbyTypePrefix: 'nacelle',
+    gatsbyNodeDefs: buildNodeDefinitions({
+      gatsbyNodeTypes: getNodeTypes(),
+      documents
+    }),
+    paginationAdapters: [PaginateNacelle]
+  };
+}
+
+async function sourceSpaceNode(config) {
+  const remoteTypeName = 'Space';
+  const document = config.gatsbyNodeDefs.get(remoteTypeName).document;
+  const result = await config.execute({
+    query: print(document),
+    operationName: `NODE_SPACE`,
+    variables: {}
+  });
+
+  await config.gatsbyApi.actions.createNode({
+    id: `${remoteTypeName}${result.data.getSpace.remoteId}`,
+    ...result.data.getSpace,
+    internal: {
+      type: `${config.gatsbyTypePrefix}${remoteTypeName}`,
+      contentDigest: config.gatsbyApi.createContentDigest(result.data.getSpace)
+    }
+  });
+}
+
+export const sourceNodes = async (gatsbyApi, pluginOptions) => {
+  const config = await createSourcingConfig(gatsbyApi, pluginOptions);
+
+  // Add explicit types to gatsby schema
+  await createSchemaCustomization(config);
+
+  // Source nodes
+  await sourceAllNodes(config);
+
+  // Source the Space node manually (as it is not sourced automatically yet):
+  await sourceSpaceNode(config);
+};
